requirements.txt
pyautogui
pillow
openai
pynput
python-dotenv
pyperclip
prompt_toolkit
